Update R/plot.R
The na.rm=TRUE option was not specified in the draw_means helper function when getting the min and max mean value for specified plot terms. As a result, an error message was received explaining that there was an unspecified ylim error when plotting, as NA was being returned as the mean min and max and then being set as the ylim lower and upper bound. This is not an issue in the draw_qnts as NA values are dealt using na.rm in the main body of the function.