Update plot.R
The plot.netsim function originally only displayed three specified plot variables when mean.col and qnts.col were not user-specified due to hardcoding of the color palette for three colors.
Combined "Set1", "Set2", and "Set3" brewer palettes for the fix to allow the max number of plotted variables to be 29.