fixes #804: warnings on `get_partners`
this PR adds a MESSAGE on `get_cumulative_edgelist` when the `el_cuml`
object do not exist yet.
If this message is printed at each time step, it means that
`update_cumulative_edgelist` is not called. Otherwise, it appears the
first time the `el_cuml`.

In addition, the `cumulative.edgelist` and `truncate.el.cuml` controls
get default values in `netsim_validate_control` to reduce the overhead
of checking for NULL each time we want to use them.

New tests are added to validate these behaviors.

This PR probably should be merged after #765 to avoid merge conflicts