On reinit, use provided params
The reinit code in EpiModel now superseeds the `reinit_msm`
however, this function behavior is to copy the params from the restart
point and not use the ones provided by netsim.

This PR reverts that as it is the main way we modify params on simple
scenarios.

Another solution would be to always use the scenario API with the
step of scenario effect greater than the restart point. But I feel like
it defeats the purpose of the restart mechanism